requirements.txt
# ── 模型训练 (recognition/) ──────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── Web 服务 (web/) ──────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.0.0
python-multipart>=0.0.9

# ── 图像预处理 (image_processing/) ──────────────────────────────────────────
opencv-python>=4.8.0
numpy>=1.24.0

# ── 训练可视化与日志 (recognition/) ─────────────────────────────────────────
matplotlib>=3.8.0
tensorboard>=2.14.0
